Refactor audio and dice bars into UI subpackages
