+++ modules/audio/ui/bar/__init__.py
@@ -1,0 +1,5 @@
+"""Audio bar UI package."""
+
+from .window import AudioBarWindow
+
+__all__ = ["AudioBarWindow"]
