+++ modules/audio/ui/bar/layout.py
@@ -1,0 +1,6 @@
+"""Layout constants for AudioBarWindow."""
+
+SEARCH_RESULTS_MENU_WIDTH = 200
+NOW_PLAYING_MENU_WIDTH = 320
+SEARCH_RESULTS_LABEL_MAX_CHARS = 25
+NOW_PLAYING_LABEL_MAX_CHARS = 35
