+++ modules/dice/ui/bar/__init__.py
@@ -1,0 +1,6 @@
+"""Dice bar UI package."""
+
+from .formatting import FormattedRoll, ResultChip, TextSegment
+from .window import DiceBarWindow
+
+__all__ = ["DiceBarWindow", "TextSegment", "ResultChip", "FormattedRoll"]
