+++ modules/dice/ui/bar/geometry.py
@@ -1,0 +1,4 @@
+"""Geometry helpers for dice bar anchoring."""
+
+INTER_BAR_GAP = 0
+HEIGHT_PADDING = 10
