+++ modules/dice/ui/bar/layout.py
@@ -1,0 +1,5 @@
+"""Layout constants for dice bar widgets."""
+
+FORMULA_ENTRY_WIDTH = 200
+ROLL_BUTTON_WIDTH = 80
+CLEAR_BUTTON_WIDTH = 70
